# Backend/api_docs.yaml
openapi: 3.0.3
info:
  title: LearnAtHome - Platform REST API (Admin, Student, Parent, Tutor)
  description: >
    APIs powering the LearnAtHome platform including Admin, Student, Parent,
    and Tutor Dashboards. Covers platform overview, scheduling, attendance,
    assignments, study materials, Q&A, doubts, earnings, and profile.
  version: "2.0.0"
servers:
  - url: http://localhost:5000
    description: Local development server
...
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Flask session cookie set by a successful POST /login as an Admin
        account. All /admin/* routes are gated by the @admin_required
        decorator, which checks session['role'] == 'Admin'.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Student with id 9999 not found."
        errors:
          nullable: true
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in (any role) and start a session
      description: >
        Validates identifier + password against Admin, Tutor, Student, and
        Parent tables (in that order). On success, clears any stale
        session, stores user_id/username/role in the Flask session cookie,
        updates last_login_at, and returns a signed JWT for API clients
        that prefer bearer-token auth over the cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, description: "username or email" }
                password: { type: string }
                remember:
                  type: boolean
                  default: false
                  description: If true, session cookie persists beyond the browser session
            example:
              identifier: "aarav.mehta@learnmail.com"
              password: "Student@123"
              remember: false
      responses:
        '200':
          description: Login successful; sets a session cookie and returns a JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Login Successful!" }
                  token: { type: string, description: "Signed JWT (HS256), 24h expiry" }
                  role: { type: string, enum: [Admin, Tutor, Student, Parent] }
                  username: { type: string, description: "Display name of the logged-in user" }
                  redirect_url: { type: string, example: "/student" }
        '400':
          description: Missing identifier/password
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Account is Pending approval or Blocked/Inactive
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/check-parent-email:
    get:
      tags: [Auth]
      summary: Look up a Parent account by email (used while filling the registration form)
      description: >
        Used by the combined Student + Parent registration form as the
        Parent Email field is filled in, so the UI can either confirm the
        existing parent account (and hide the "new parent" fields) or
        reveal the fields needed to create a new Parent. This is a
        best-effort UI hint only; POST /auth/register performs its own
        authoritative lookup in the same transaction as account creation.
      security: []
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
      responses:
        '200':
          description: Lookup completed (whether or not a match was found)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  exists: { type: boolean }
                  parent:
                    type: object
                    nullable: true
                    description: Present only when exists=true
                    properties:
                      parent_id: { type: integer }
                      parent_name: { type: string }
                      email: { type: string }
        '400':
          description: Missing or invalid email
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new Student (with linked Parent), or a new Tutor account
      description: >
        Self-service sign-up. Supports two request shapes:

        1. **Combined Student + Parent** (current registration form) - body
        contains a nested `student` object and a `parent` object. The
        Parent's email is looked up first: if a Parent with that email
        already exists, the new Student is linked to it via
        `student.parent_id` and no new Parent row is created; otherwise a
        new Parent is created from the supplied `parent` fields and the
        Student is linked to its new `parent_id`. Parent creation (when
        needed) and Student creation happen in a single DB transaction, so
        a Student is never left pointing at an invalid/partially-created
        Parent.

        2. **Legacy flat single-role body** (`name`/`email`/`password`/
        `confirm_password`/`role`) - now only accepts `role: "Tutor"`.
        Standalone Parent registration and flat (unlinked) Student
        registration have been removed; Student accounts must go through
        the combined shape above so every Student is linked to a Parent.

        In both cases, if the caller already has an active session, the
        request is rejected with 409 Conflict (no account is created) and
        the response tells them which account they're already logged in
        as, along with a redirect_url for that existing role.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [student, parent]
                  properties:
                    student:
                      type: object
                      required: [name, email, password, confirm_password, subject_ids]
                      properties:
                        name: { type: string, minLength: 2 }
                        email: { type: string, format: email }
                        password: { type: string, minLength: 6 }
                        confirm_password: { type: string, minLength: 6 }
                        phone_no: { type: string, nullable: true }
                        school: { type: string, nullable: true, description: "Optional" }
                        subject_ids:
                          type: array
                          minItems: 1
                          description: "Required - at least one subject must be selected"
                          items: { type: integer }
                    parent:
                      type: object
                      required: [email]
                      description: >
                        Only `email` is required if the Parent already
                        exists. `name`/`password`/`confirm_password` are
                        required when the Parent email is new (i.e. no
                        existing Parent account was found for it).
                      properties:
                        email: { type: string, format: email }
                        name: { type: string, minLength: 2 }
                        password: { type: string, minLength: 6 }
                        confirm_password: { type: string, minLength: 6 }
                        phone_no: { type: string, nullable: true }
                - type: object
                  required: [name, email, password, confirm_password, role]
                  properties:
                    name: { type: string, minLength: 2, description: "Full name" }
                    email: { type: string, format: email }
                    password: { type: string, minLength: 6 }
                    confirm_password: { type: string, minLength: 6 }
                    role:
                      type: string
                      enum: [Tutor]
            examples:
              student_with_existing_parent:
                summary: Student registration, Parent email already exists
                value:
                  student:
                    name: "New Student"
                    email: "new.student@learnmail.com"
                    password: "Passw0rd!"
                    confirm_password: "Passw0rd!"
                  parent:
                    email: "ravi@gmail.com"
              student_with_new_parent:
                summary: Student registration, Parent email is new
                value:
                  student:
                    name: "New Student"
                    email: "new.student2@learnmail.com"
                    password: "Passw0rd!"
                    confirm_password: "Passw0rd!"
                  parent:
                    email: "new.parent@parentmail.com"
                    name: "New Parent"
                    password: "ParentPass1!"
                    confirm_password: "ParentPass1!"
                    phone_no: "+1 200 555 2000"
              tutor_legacy_flat:
                summary: Tutor registration (legacy flat shape)
                value:
                  name: "New Tutor"
                  email: "new.tutor@learnmail.com"
                  password: "Passw0rd!"
                  confirm_password: "Passw0rd!"
                  role: "Tutor"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Registration Successful! Please log in." }
                  parent:
                    type: object
                    nullable: true
                    description: Only present for the combined Student + Parent shape
                    properties:
                      parent_id: { type: integer }
                      parent_name: { type: string }
                      email: { type: string }
                      newly_created:
                        type: boolean
                        description: true if a new Parent row was created, false if an existing one was reused
        '400':
          description: >
            Validation failed - e.g. name too short, invalid email,
            password too short, password/confirm_password mismatch,
            Student or Parent email already registered, missing Parent
            details when the Parent email is new, or role not in [Tutor]
            for the legacy flat shape
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: >
            Caller already has an active session (logged in as a
            different, or the same, account) - no account is created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  already_logged_in: { type: boolean, example: true }
                  message: { type: string, example: "You're already logged in as Aarav Mehta (Student). Please log out before creating a new account." }
                  role: { type: string }
                  redirect_url: { type: string, example: "/student" }
        '500':
          description: Unexpected error while creating the account(s) (transaction rolled back)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...